improved test speed with encoder placeholders
