+++ tests/demo_encoders/__init__.py
@@ -1,0 +1,2 @@
+from .classic_net import ClassicNet
+from .cursed_net import CursedNet
