+++ tests/demo_encoders/cursed_net.py
